requirements.txt
ascii-magic==2.3.0
cachetools==5.3.0
colorama==0.4.6
google-auth==2.17.0
google-auth-oauthlib==1.0.0
gspread==5.7.2
numpy==1.24.2
oauthlib==3.2.2
Pillow==9.4.0
prettytable==3.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2023.3
requests-oauthlib==1.3.1
rsa==4.9
